+++ src/components/Header.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react';
 import { Bell, Search, MapPin } from 'lucide-react';
+import useKakaoLoader from '../hooks/useKakaoLoader'; // 공통 훅 사용
 
 interface KakaoAddress {
   address_name: string;
@@ -34,77 +35,48 @@ interface KakaoGeocoderResult {
   };
 }
 
-declare global {
-  interface Window {
-    kakao: any;
-  }
-}
-
 export default function Header() {
   const [dongName, setDongName] = useState('로딩 중...');
+  const isKakaoLoaded = useKakaoLoader(); // 공통 훅 사용
 
   useEffect(() => {
-    const kakaoKey = import.meta.env.VITE_KAKAO_API;
-    if (!kakaoKey) {
-      console.error('Kakao API Key가 비어 있음');
-      setDongName('API 키 누락');
+    if (!isKakaoLoaded) return; // 안전 guard
+
+    if (!navigator.geolocation) {
+      setDongName('위치 정보 지원 불가');
       return;
     }
 
-    // SDK가 이미 로드되었는지 확인
-    const existingScript = document.querySelector(`script[src*="dapi.kakao.com"]`);
-    if (existingScript) {
-      // 이미 로드된 경우 maps.load만 호출
-      loadKakaoMap();
-    } else {
-      // SDK 동적 삽입
-      const script = document.createElement('script');
-      script.src = `https://dapi.kakao.com/v2/maps/sdk.js?appkey=${kakaoKey}&autoload=false&libraries=services`;
-      script.async = true;
-      script.onload = loadKakaoMap;
-      document.head.appendChild(script);
-    }
+    navigator.geolocation.getCurrentPosition(
+      (position) => {
+        const { latitude, longitude } = position.coords;
 
-    function loadKakaoMap() {
-      window.kakao.maps.load(() => {
-        if (!navigator.geolocation) {
-          setDongName('위치 정보 지원 불가');
-          return;
-        }
+        const geocoder = new window.kakao.maps.services.Geocoder();
 
-        navigator.geolocation.getCurrentPosition(
-          (position) => {
-            const { latitude, longitude } = position.coords;
-
-            // ✅ SDK가 완전히 로드된 후에만 사용 가능
-            const geocoder = new window.kakao.maps.services.Geocoder();
-
-            geocoder.coord2Address(
-              longitude,
-              latitude,
-              (result: KakaoGeocoderResult[], status: string) => {
-                if (status === window.kakao.maps.services.Status.OK) {
-                  const address = result[0]?.address;
-                  if (address?.region_3depth_name) {
-                    setDongName(address.region_3depth_name);
-                  } else {
-                    setDongName('주소 정보 없음');
-                  }
-                } else {
-                  console.warn('역지오코딩 실패:', status);
-                  setDongName('주소 조회 실패');
-                }
-              },
-            );
-          },
-          (error) => {
-            console.error('위치 권한 오류:', error);
-            setDongName('위치 권한 거부됨');
+        geocoder.coord2Address(
+          longitude,
+          latitude,
+          (result: KakaoGeocoderResult[], status: string) => {
+            if (status === window.kakao.maps.services.Status.OK) {
+              const address = result[0]?.address;
+              if (address?.region_3depth_name) {
+                setDongName(address.region_3depth_name);
+              } else {
+                setDongName('주소 정보 없음');
+              }
+            } else {
+              console.warn('역지오코딩 실패:', status);
+              setDongName('주소 조회 실패');
+            }
           },
         );
-      });
-    }
-  }, []);
+      },
+      (error) => {
+        console.error('위치 권한 오류:', error);
+        setDongName('위치 권한 거부됨');
+      },
+    );
+  }, [isKakaoLoaded]); // Kakao가 안전히 로드된 후 실행
 
   return (
     <header className="w-full pt-[60px] bg-white shadow fixed pb-2 z-50">
